+++ packages/mini-apps-ui-kit-react/src/components/Haptic/index.ts
@@ -1,0 +1,2 @@
+export { Haptic } from "./Haptic";
+export { useHaptics } from "./use-haptics";