+++ packages/mini-apps-ui-kit-react/src/index.ts
@@ -53,6 +53,7 @@ export { ToggleGroupRoot, ToggleGroupItem } from "./components/ToggleGroup";
 export { Tabs, TabItem } from "./components/Tabs";
 export { LiveFeedback } from "./components/LiveFeedback";
 export { Skeleton, SkeletonTypography } from "./components/Skeleton";
+export { Haptic, useHaptics } from "./components/Haptic";
 
 // Export the Tailwind plugin
 export { default as uiKitTailwindPlugin } from "./tailwind";